+++ src/Control/Eff/Concurrent/Api/Client.hs
@@ -5,6 +5,7 @@ module Control.Eff.Concurrent.Api.Client
   ( -- * Calling APIs directly
     cast
   , call
+  , callWithTimeout
   -- * Server Process Registration
   , castRegistered
   , callRegistered
@@ -20,6 +21,8 @@ import           Control.Eff.Reader.Strict
 import           Control.Eff.Concurrent.Api
 import           Control.Eff.Concurrent.Api.Request
 import           Control.Eff.Concurrent.Process
+import           Control.Eff.Concurrent.Process.Timer
+import           Control.Eff.Log
 import           Data.Typeable                  ( Typeable )
 import           Control.DeepSeq
 import           GHC.Stack
@@ -48,6 +51,8 @@ cast (Server pid) castMsg = sendMessage pid (Cast castMsg)
 --
 -- The type signature enforces that the corresponding 'Api' clause is
 -- 'Synchronous'.
+--
+-- __Always prefer 'callWithTimeout' over 'call'__
 call
   :: forall result api r q
    . ( SetMember Process (Process q) r
@@ -64,8 +69,8 @@ call
   -> Api api ( 'Synchronous result)
   -> Eff r result
 call (Server pidInternal) req = do
-  fromPid <- self
   callRef <- makeReference
+  fromPid <- self
   let requestMessage = Call callRef fromPid $! req
   sendMessage pidInternal requestMessage
   let selectResult :: MessageSelector result
@@ -77,6 +82,64 @@ call (Server pidInternal) req = do
         in  selectMessageWith extractResult
   resultOrError <- receiveWithMonitor pidInternal selectResult
   either (interrupt . becauseProcessIsDown) return resultOrError
+
+
+-- | Send an 'Api' request and wait for the server to return a result value.
+--
+-- The type signature enforces that the corresponding 'Api' clause is
+-- 'Synchronous'.
+--
+-- If the server that was called dies, this function interrupts the
+-- process with 'ProcessDown'.
+-- If the server takes longer to reply than the given timeout, this
+-- function interrupts the process with 'ProcessTimeout'.
+--
+-- __Always prefer this function over 'call'__
+--
+-- @since 0.22.0
+callWithTimeout
+  :: forall result api r q
+   . ( SetMember Process (Process q) r
+     , Member Interrupts r
+     , Typeable api
+     , Typeable (Api api ( 'Synchronous result))
+     , NFData (Api api ( 'Synchronous result))
+     , Typeable result
+     , NFData result
+     , Show result
+     , Member Logs r
+     , Lifted IO q
+     , Lifted IO r
+     , HasCallStack
+     )
+  => Server api
+  -> Api api ( 'Synchronous result)
+  -> Timeout
+  -> Eff r result
+callWithTimeout serverP@(Server pidInternal) req timeOut = do
+  fromPid <- self
+  callRef <- makeReference
+  let requestMessage = Call callRef fromPid $! req
+  sendMessage pidInternal requestMessage
+  let selectResult =
+        let extractResult
+              :: Reply (Api api ( 'Synchronous result)) -> Maybe result
+            extractResult (Reply _pxResult callRefMsg result) =
+              if callRefMsg == callRef then Just result else Nothing
+        in selectMessageWith extractResult
+  resultOrError <- receiveSelectedWithMonitorAfter pidInternal selectResult timeOut
+  let onTimeout timerRef = do
+        let msg = "call timed out after "
+                  ++ show timeOut ++ " to server: "
+                  ++ show serverP ++ " from "
+                  ++ show fromPid ++ " "
+                  ++ show timerRef
+        logWarning' msg
+        interrupt (ProcessTimeout msg)
+      onProcDown p = do
+        logWarning' ("call to dead server: "++ show serverP ++ " from " ++ show fromPid)
+        interrupt (becauseProcessIsDown p)
+  either (either onProcDown onTimeout) return resultOrError
 
 -- | Instead of passing around a 'Server' value and passing to functions like
 -- 'cast' or 'call', a 'Server' can provided by a 'Reader' effect, if there is